Move the in-memory store to elide-core
