+++ elide-core/src/main/java/com/yahoo/elide/core/datastore/inmemory/Operation.java
@@ -1,9 +1,9 @@
 /*
- * Copyright 2015, Yahoo Inc.
+ * Copyright 2017, Yahoo Inc.
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
-package com.yahoo.elide.datastores.inmemory;
+package com.yahoo.elide.core.datastore.inmemory;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
